# config/swagger.json
{
    "swagger": "2.0",
    "host": "localhost:8000",
    "schemes": [
        "http"
    ],
    "info": {
        "title": "FLOW CARDIT API",
        "version": "1.0"
    },
    "paths": {
...
        "/cards/{cardId}/decks": {
            "get": {
                "description": "Show decks by cardId",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show decks by cardId"
                    }
                }
            },
            "post": {
                "description": "Create cardDeck",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "body",
                        "required": "true",
                        "schema": {
                            "properties": {
                                "deckId": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                ]
            }
        },
        "/cards/{cardId}/decks/{deckId}": {
            "put": {
                "description": "Update cardDeck",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "body",
                        "required": "true",
                        "schema": {
                            "$ref": "#/definitions/CardDeckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update successful"
                    },
                    "400": {
                        "description": "CardDeck with given id does not exist"
                    }
                }
            },
            "delete": {
                "description": "Delete cardDeck",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successful"
                    },
                    "400": {
                        "description": "CardDeck with given id does not exist"
                    }
                }
            }
        },
        "/groups": {
            "get": {
                "description": "Index of groups",
                "responses": {
                    "200": {
                        "description": "Returns a list of groups"
                    }
                }
            },
            "post": {
                "description": "Creates new group",
                "parameters": [
                    {
                        "$ref": "#/parameters/groupData"
                    },
                    {
                        "name": "deckId",
                        "in": "body",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New group created"
                    },
                    "500": {
                        "description": "Failed to create group"
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "description": "Show group by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns group by id",
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    },
                    "400": {
                        "description": "Group with given id does not exist"
                    }
                }
            },
            "put": {
                "description": "Update group by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    },
                    {
                        "$ref": "#/parameters/groupData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group updated",
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    },
                    "400": {
                        "description": "Failed to update group"
                    }
                }
            },
            "delete": {
                "description": "Delete group by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group deleted",
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    },
                    "400": {
                        "description": "Group with given id does not exist."
                    }
                }
            }
        },
...
        "/groups/{groupId}/decks": {
            "get": {
                "description": "Show decks by groupId",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show decks by groupId"
                    }
                }
            },
            "post": {
                "description": "Create groupDeck",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "body",
                        "required": "true",
                        "schema": {
                            "properties": {
                                "deckId": {
                                    "type": "integer"                                    
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Creating successful"
                    }
                }
            }
        },
        "/groups/{groupId}/decks/{deckId}": {
            "put": {
                "description": "Update deck_group by groupId",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "body",
                        "required": "true",
                        "schema": {
                            "properties": {
                                "deckId": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update successful"
                    }
                }
            },
            "delete": {
                "description": "Delete deck by id",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    },
                    {
                        "name": "deckId",
                        "in": "path",
                        "required": "true",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successful"
                    }
                }
            }
        },
        "/decks": {
            "get": {
                "description": "Index of decks",
                "responses": {
                    "200": {
                        "description": "Returns a list of decks"
                    },
                    "400": {
                        "description": "Failed to show decks"
                    }
                }
            },
            "post": {
                "description": "Creates new deck",
                "parameters": [
                    {
                        "$ref": "#/parameters/deckData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New deck created"
                    },
                    "500": {
                        "description": "Failed to create deck"
                    }
                }
            }
        },
        "/decks/{id}": {
            "get": {
                "description": "Show deck by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns deck by id",
                        "schema": {
                            "$ref": "#/definitions/Deck"
                        }
                    },
                    "400": {
                        "description": "Deck with given id does not exist"
                    }
                }
            },
            "put": {
                "description": "Update deck by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    },
                    {
                        "$ref": "#/parameters/deckData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deck updated",
                        "schema": {
                            "$ref": "#/definitions/Deck"
                        }
                    },
                    "400": {
                        "description": "Failed to update deck"
                    }
                }
            },
            "delete": {
                "description": "Delete deck by id",
                "parameters": [
                    {
                        "$ref": "#/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deck deleted",
                        "schema": {
                            "$ref": "#/definitions/Deck"
                        }
                    },
                    "400": {
                        "description": "Deck with given id does not exist."
                    }
                }
            }
        },
...
        "/decks/{deckId}/groups": {
            "get": {
                "description": "Show groups by deckId",
                "parameters": [
                    {
                        "$ref": "#/parameters/deckId"
                    }
                ]
            }
        }
    },
    "definitions": {
...
        "CardDeckRequest": {
            "properties": {
                "deckId": {
                    "type": "integer"
                }
            }
        },
...
        "Deck": {
            "properties": {
                "subject": {
                    "type": "string",
                    "allowNull": false,
                    "notEmpty": true,
                    "unique": true,
                    "validate": {
                        "len": [
                            3,
                            500
                        ]
                    }
                }
            }
        },
...
        "Group": {
            "properties": {
                "name": {
                    "type": "string",
                    "allowNull": false,
                    "notEmpty": true,
                    "unique": true,
                    "validate": {
                        "len": [
                            3,
                            500
                        ]
                    }
                }
            }
        },
...
    },
    "parameters": {
        "id": {
            "name": "id",
            "in": "path",
            "required": "true",
            "type": "integer"
        },
        "deckId": {
            "name": "deckId",
            "in": "path",
            "required": "true",
            "type": "integer"
        },
...
        "groupData": {
            "name": "group",
            "in": "body",
            "schema": {
                "$ref": "#/definitions/Group"
            }
        },
        "deckData": {
            "name": "deck",
            "in": "body",
            "schema": {
                "$ref": "#/definitions/Deck"
            }
        }
    }
}